- activated compiler warnings for auto(un)boxing - fixed a compiler warning and changed compiler settings for Samples project
git-svn-id: http://mapsforge.googlecode.com/svn/trunk@1436 a424f762-2dd8-11df-bfeb-7b4c664f286a

+++ mapsforge/src/org/mapsforge/android/maps/theme/Circle.java
@@ -43,7 +43,7 @@ final class Circle extends RenderInstruction {
 			String value = attributes.getValue(i);
 
 			if ("r".equals(name)) {
-				radius = Float.parseFloat(value);
+				radius = Float.valueOf(Float.parseFloat(value));
 			} else if ("scale-radius".equals(name)) {
 				scaleRadius = Boolean.parseBoolean(value);
 			} else if ("fill".equals(name)) {
